refactor(chatgpt): ChatGPT Assistant Client와 Basic ChatGPT Client를분리
구 버전의 코드를 chatgpt_assistant package에 분리하고
ChatClient를 활용하는 신 버전의 코드를 chatgpt package에 추가할 예정.
