+++ src/test/java/com/newbini/quizard/client/AssistantGeneratorTest.java
@@ -1,10 +1,6 @@
 package com.newbini.quizard.client;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import com.newbini.quizard.client.chatgpt.assistant.AssistantGenerator;
-import com.newbini.quizard.dto.response.AssistantObject;
-import com.newbini.quizard.dto.response.ThreadObject;
-import com.newbini.quizard.dto.response.VectorStoreObject;
 import org.apache.tomcat.util.json.ParseException;
 import org.junit.jupiter.api.Test;
