+++ src/test/java/com/newbini/quizard/client/DeleteAssistantsTest.java
@@ -1,6 +1,5 @@
 package com.newbini.quizard.client;
 
-import com.newbini.quizard.dto.response.ListAssistantsObject;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.web.client.TestRestTemplate;
