+++ src/test/java/com/newbini/quizard/client/MessageGeneratorTest.java
@@ -1,9 +1,6 @@
 package com.newbini.quizard.client;
 
-import com.newbini.quizard.client.chatgpt.assistant.AssistantGenerator;
 import com.newbini.quizard.client.chatgpt.message.MessageGenerator;
-import com.newbini.quizard.dto.response.MessageObject;
-import com.newbini.quizard.dto.response.ThreadObject;
 import org.junit.jupiter.api.Test;
 
 import java.io.File;
